requirements.txt
ccxt>=4.4.0,<5.0.0
pytest>=8.0.0,<9.0.0
python-dotenv>=1.0.0,<2.0.0
websockets>=12.0,<16.0
tzdata>=2024.1
